+++ src/imgs-preview/index.js
@@ -2,6 +2,6 @@ import { Func } from './core';
 
 import ImgsPreview from './imgs-preview';
 
-ImgsPreview.popup = Func.popup;
+ImgsPreview.open = Func.popup;
 
 export default ImgsPreview;