+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### 1.3.16
 - picker: 新增组件picker-view, picker-popup(用于自定义, date-picker-core也是基于此开发的)
 - date-picker: 新增组件date-picker-view
+- picker: `MPicker.popup` -> `MPicker.open`
+- date-picker: `MDatePicker.popup` -> `MDatePicker.open`
+- imgs-preview: `ImgsPreview.popup` -> `ImgsPreview.open`
 
 ### 1.3.8
 - popup: `position` -> `placement`, 新增theme(light/dark/none)